Ajusta metodo excluir
Cria tratativa para excluir os seus produtos

+++ system/Modules/Listagem/app/Http/Controllers/ListagemController.php
@@ -5,8 +5,10 @@ namespace Modules\Listagem\Http\Controllers;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Inertia\Inertia;
 use Modules\Listagem\Models\Lista;
+use Modules\Listagem\Models\Produto;
 
 class ListagemController extends Controller
 {
@@ -60,7 +62,29 @@ class ListagemController extends Controller
     public function excluir($id)
     {
         $lista = Lista::find($id);
-        $lista->delete();
-        return to_route('listagem.index');
+
+        if(!$lista) {
+            return redirect()->back()->withErrors('Lista não encontrada.');
+        }
+
+        DB::beginTransaction();
+
+        try {
+            $produtos = Produto::where('lista_id', $id)->get();
+            foreach ($produtos as $produto) {
+                $produto->delete();
+            }
+
+            $lista->delete();
+
+            DB::commit();
+
+            return to_route('listagem.index')->with('success', 'Lista e produtos excluídos com sucesso.');
+        } catch (\Exception $e) {
+            DB::rollBack();
+
+            return redirect()->back()->withErrors('Erro ao tentar excluir a lista.');
+        }
     }
+
 }
